+++ data/src/main/kotlin/lmm/moneylog/data/transaction/model/Transaction.kt
@@ -3,7 +3,7 @@ package lmm.moneylog.data.transaction.model
 import lmm.moneylog.data.time.model.DomainTime
 
 data class Transaction(
-    val id: Int = -1,
+    val id: Long = -1L,
     val value: Double,
     val description: String,
     val date: DomainTime,
