+++ data/src/main/kotlin/lmm/moneylog/data/transaction/repositories/impls/DeleteTransactionRepositoryImpl.kt
@@ -6,7 +6,7 @@ import lmm.moneylog.data.transaction.repositories.interfaces.DeleteTransactionRe
 class DeleteTransactionRepositoryImpl(
     private val transactionDao: TransactionDao
 ) : DeleteTransactionRepository {
-    override suspend fun delete(id: Int) {
+    override suspend fun delete(id: Long) {
         transactionDao.delete(id)
     }
 }
